Added Historical (Unweighted) AI Node and enabling code.
Added an AI node called HistoricalUnweightedAINodeMulti. This is currently storing historical points for each object within each camera in order to calculate a mean velocity (the quantity of movement). Also added a scenario which uses this camera.

Enabled the historical AI node in SimCore by adding a couple of else-if cases, and made the historical scenario the default one in run.sh.

Mild changes to comp.sh to use non-quoted strings.