Removed unused add_camera method
Removed the additional add_camera method which allowed the passing of an
integer 'aiID' argument to specify the type of camera to be added. This
was not being used but, more importantly, it was nearly identical code
to the other add_camera method and would likely give rise to bugs. If
it's needed again, it can be re-written quickly and more safely.