Fixed bounce angle bug.
When a travelling object in the simulator hits a wall, it bounces back
by means of manipulation of the object's heading. This was sending
objects in strange directions due to the heading being modified in
degrees whereas it was being read in radians. Fixed by adding a
'getTurnaroundAngle()' method which computes the angle correctly in
radians.