Fixed a bug allowing scenarios with identically named cameras
If you specify a scenario with multiple cameras with no name, they are
all added with the empty string "" as their name. This doesn't cause an
exception but causes handover code to fail due to its reliance on the
name string for identification. This fix throws an exception when more
than one camera is added without a name, or multiple cameras have the
same name.