New scenario files from demo
Took the demo scenarios embedded in the code (within
SimCoreModel.loadDemo(int) ) and created three scenario files from them
for the HistoricalUnweighted AI node. The third currently has no objects
because the demo specified waypoints instead, but objects can be easily
added with the click of the Obj++ button in the main window.