Optimised and generalised compile and run scripts
Compile script no longer packages the entire library JAR. Instead, it has been added to the classpath of the built JAR. This does mean that the library JAR must be present separately before CamSim will run. It also means it has to be in the same directory or in a directory on the classpath.

I have fixed an issue where the compilation would not recognise unreferenced classes (i.e. those referenced only at runtime using reflection). These are now automatically included as long as they match src/epics/ai/*.java