Added solutions to overstay problem and instantiation problem
Instantiation problem: Never seen an object before so we have no
avg TS to calculate history based bids. Set a default for this
(1.0 for now but should investigate).

Overstay problem: AvgTS is e.g. 10 and object is around for 15 TS. What
do we bid? We can't use our current formula so we use a default
bid coefficient instead