Fixed CameraController constructor bug. Fixed determinism/hashcode
- CameraController constructor was taking in arguments and assigning
them to private static fields! Outright bug which is now fixed (they are
now instance fields)

- Previously, the code was not deterministic for the same AI node. This
was not trivial to find. The reason was that tracked objects were being
stored in a hashmap, which was using hashes of CameraControllers as
keys. These were being iterated over in the order of keys in the map,
which changed during every runtime because it was using the default
Object.hashCode() method which only guarantees constancy during each
JVM run. I fixed it to use a basic hashCode method, so now
CameraControllers are iterated over in a deterministic way and several
runs will obtain the same solution.