Added summaryFile generation and enabling code. Minor formatting.
Instead of simply relying on output.csv files which tell us about
individual timesteps, there is now a summary file that can be generated
using the --summaryfile or -f arguments on the command line. These
currently include the same information as the output files plus
cumulative data such as cumulative utility. The idea of this is to sum
up a whole experiment in a one-line summary to compare the fitness of
different algorithms/configurations.

Also removed some redundant code and cleaned up some formatting.