Created experiment run script. Improved run script and gitignore
- Created exp.sh, which runs many experiments and helpfully places all
the resulting .csv files, summary files and standard out to a specified
directory. Experiments are run with a different seed every time, which
is currently a number from 1 up to the number of experiments (e.g.
1-30)
- Run script now checks for existence of the build.jar and library jar,
as well as now allowing the specification of arguments, which are
then passed straight to the JVM
- GitIgnore file updated to ignore files ending in ~ and files beginning
and ending with # (Linux and Emacs temp files)