Restructured HistoricalAINode into active/passive. Added scenarios
- Major restructuring of the Historical node. The idea is to have an
active and a passive historical node, but since these would have to
extend different classes, there would have to be duplicated code. The
solution implemented is a little bit ugly but there is minimal code
duplication. Essentially, there is a HistoricalAINodeActive and Passive,
each of which is a shell for the HistoricalNodeHelper which has the real
logic and state inside. The Historical nodes (shell classes) are
duplicates so any code changed in one must be changed in the other,
unfortunately. Please forgive me.
- Added two new scenarios to show off the advantages of history-based
bidding
- Updated previous scenarios due to new names for historical nodes