Added history-per-category. Cleaned up hist code
- Added history per category (and lots of supporting code). That is, the
historical information kept for each camera is now split according to
the category the object was classified into. This can be enabled and
disabled by the HistPerCategoryEnabled param (which is only effective in
conjunction with the ClassificationEnabled param). It is turned on by
default
- Cleaned up constructors of hist nodes. Previously they would take
values for PreInstantiationBidCoefficient and OverstayBidCoefficient,
but now these values are set exclusively by means of the RunParams
mechanism. This simplifies the code as well as the usage and makes the
whole thing easier to maintain.
- Some formatting and println changes (in particular, removing too many
decimal places)