Added scenarios 7,8,9 and classification experiment scripts
- Added scenarios 7, 8 and 9, each with an 'a' and 'b' version. These
are designed to show the behaviour of classification to varying degrees.
The 'a' version of each has one object, whereas the 'b' version has
multiple objects. The 'b' version is the practically useful one, while
the 'a' simply exists for comparison
- Added HistEnabled param for HistNode. Turning this on will make Hist
behave like its parent (so Active or Passive depending on the class
chosen). This doesn't have any practical use, but is instead for testing
purposes. Of course, hist is enabled by default
- Changed default PreInst and Overstay values for Hist to match
experiment results
- Added ClassyParams.properties, a params file that makes use of
classification and histpercategory
- Added 'exp-class.sh', a script to run experiments across different
classication parameters for the experiments on classification
- Added 'exp-results-class.sh', a script to gather results from the
exp-class.sh script