Fixed two major classification bugs
- Fixed a major bug with classification, where itro objects were being
stored in the pheromone maps, rather than the categories. The compiler
didn't pick this up due to HashMap taking 'Object' as a param. We are
now passing in category rather than the itro where required
- Fixed bug with classification in the HistAINode. When an object was
lost, its category was not retrievable because its historical points
would have been erased. This is now being stored in a separate HashMap
called lastSeenCategoryForObj, which is recorded whenever an object
leaves the field of view for future reference, and erased once it
re-enters
- Added some handling for when category == 0 (i.e. no idea what category
because not enough info)