Removed code which removed tracked objects when they became invisible
Previously, when a tracked object became invisible, it would also be removed from the tracked objects list. However, the tracked objects list should be used to store objects for which a camera has tracking responsibility. The checkIfTracedGotLost() method has now been removed, along with the removeTracedObject() call within the removeVisibleObject() method.

The result of this is that we have communication during time steps when objects are not visible (as desired).