Moved 'lastConfidence' code into Passive AI node
Both the ActiveAINode and PassiveAINode contain code which refers to 'lastConfidence', i.e. the confidence value in the previous time step for some objects. However, this takes effect exclusively in the Passive node, so I have moved it all to the Passive node. As an added bonus, the checkConfidences() code in Active is now only three lines long.