CheckBidsForObjects() bug removed from Passive node
The Passive node was overriding the CheckBidsForObjects() code from Active, with only a slight variation in the code. It was calling getConfidence() rather than calculateValue(), which is an outright bug. It affects code where calculateValue() != getConfidence(). Since this was the only difference, the whole method has been removed from Passive